New OpenZFS refactor - factor platform specific bits out of libzfs
