+++ lib/libzfs/Makefile.am
@@ -27,6 +27,15 @@ USER_C = \
 	libzfs_status.c \
 	libzfs_util.c
 
+if BUILD_LINUX
+USER_C += \
+	os/linux/libzfs_mount_os.c \
+	os/linux/libzfs_pool_os.c \
+	os/linux/libzfs_sendrecv_os.c \
+	os/linux/libzfs_util_os.c
+endif
+
+
 KERNEL_C = \
 	algs/sha2/sha2.c \
 	zfeature_common.c \
@@ -51,13 +60,18 @@ nodist_libzfs_la_SOURCES = \
 
 libzfs_la_LIBADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
-	$(top_builddir)/lib/libshare/libshare.la \
 	$(top_builddir)/lib/libuutil/libuutil.la \
 	$(top_builddir)/lib/libzfs_core/libzfs_core.la \
 	$(top_builddir)/lib/libzutil/libzutil.la
 
-libzfs_la_LIBADD += -lm $(LIBSSL)
+if BUILD_LINUX
+libzfs_la_LIBADD += \
+	$(top_builddir)/lib/libshare/libshare.la
+endif
+
 libzfs_la_LDFLAGS = -version-info 2:0:0
+
+libzfs_la_LIBADD += -lm $(LIBSSL)
 
 EXTRA_DIST = $(libzfs_pc_DATA) $(USER_C)
